Assign subfigures to Fig. 10
The two plots in Fig. 10 were referenced separately in the text, but were appearing as Fig. 6.2 (the section containing the reference). The subcaption package has been added to give a/b subfigure lettering, and the caption package to set the letters not to display. \captionsetup can be commented out to have the letters appear.

This does have the unfortunate effect of (slightly) increasing the space between the figures and the caption.