Update iphas-dr2.tex
I've been through everything now, except the figure captions (which I probably will leave to you).  Note that I've numbered the Sh 2-82 candidate Ha emitters but don't know which two are the objects in common with Yu+Wang (over to you to resolve the 'x and y' now in the text).  In section 5, you'll find a couple of comments, that begin with GEERT that you need to address.  Also you now have a Parker05 reference to add to the bibliography.  ...still to do: expand section 8.